requirements.txt
pandas>=2.0.0
numpy>=1.24.0
ollama>=0.1.0
python-dateutil>=2.8.0
faker>=18.0.0
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0
PyYAML>=6.0